feat: domain 분리
