+++ src/index.js
@@ -1,6 +1,7 @@
+import { GameModel } from './gameModel.js';
 import { GameProcessor } from './gameProcessor.js';
 import { GameView } from './GameView.js';
 
-const gameProcessor = new GameProcessor(new GameView());
+const gameProcessor = new GameProcessor(new GameModel(), new GameView());
 
 gameProcessor.startGame();
